dev-requirements.txt
ConfigArgParse
coloredlogs
pyzmq
aiozmq
aiohttp>=1.0
aioredis>=0.2.8
namedlist
sqlalchemy
aiopg
python-dateutil>=2.5
simplejson
uvloop>=0.5.4
-e git+ssh://git@github.com/lablup/sorna-common#egg=sorna-common
pytest
pytest-mock
flake8
pep8-naming
-e .
